+++ client/app/reader/DocumentsTable.jsx
@@ -14,130 +14,19 @@ import * as Constants from './constants';
 import CommentIndicator from './CommentIndicator';
 import DropdownFilter from './DropdownFilter';
 import { bindActionCreators } from 'redux';
-import Link from '../components/Link';
 import Highlight from '../components/Highlight';
 import { setDocListScrollPosition, changeSortState, clearTagFilters, clearCategoryFilters,
-  setTagFilter, setCategoryFilter, selectCurrentPdfLocally, toggleDropdownFilterVisibility } from './actions';
+  setTagFilter, setCategoryFilter, toggleDropdownFilterVisibility } from './actions';
 import { getAnnotationsPerDocument } from './selectors';
 import {
-  SelectedFilterIcon, UnselectedFilterIcon, rightTriangle,
   SortArrowUp, SortArrowDown, DoubleArrow } from '../components/RenderFunctions';
 import DocCategoryPicker from './DocCategoryPicker';
 import DocTagPicker from './DocTagPicker';
+import FilterIcon from './FilterIcon';
+import LastReadIndicator from './LastReadIndicator';
+import DocTypeColumn from './DocTypeColumn';
 
 const NUMBER_OF_COLUMNS = 6;
-
-class FilterIcon extends React.PureComponent {
-  render() {
-    const {
-      handleActivate, label, getRef, selected, idPrefix
-    } = this.props;
-
-    const handleKeyDown = (event) => {
-      if (event.key === ' ' || event.key === 'Enter') {
-        handleActivate(event);
-        event.preventDefault();
-      }
-    };
-
-    const className = 'table-icon';
-
-    const props = {
-      role: 'button',
-      getRef,
-      'aria-label': label,
-      className,
-      tabIndex: '0',
-      onKeyDown: handleKeyDown,
-      onClick: handleActivate
-    };
-
-    if (selected) {
-      return <SelectedFilterIcon {...props} idPrefix={idPrefix} />;
-    }
-
-    return <UnselectedFilterIcon {...props} />;
-  }
-}
-
-FilterIcon.propTypes = {
-  label: PropTypes.string.isRequired,
-  iconName: PropTypes.string,
-  handleActivate: PropTypes.func,
-  getRef: PropTypes.func,
-  idPrefix: PropTypes.string.isRequired,
-  className: PropTypes.string
-};
-
-class LastReadIndicator extends React.PureComponent {
-  render() {
-    if (!this.props.shouldShow) {
-      return null;
-    }
-
-    return <span
-      id="read-indicator"
-      ref={this.props.getRef}
-      aria-label="Most recently read document indicator">
-      {rightTriangle()}
-    </span>;
-  }
-}
-
-const lastReadIndicatorMapStateToProps = (state, ownProps) => ({
-  shouldShow: state.readerReducer.ui.pdfList.lastReadDocId === ownProps.docId
-});
-const ConnectedLastReadIndicator = connect(lastReadIndicatorMapStateToProps)(LastReadIndicator);
-
-class DocTypeColumn extends React.PureComponent {
-  boldUnreadContent = (content, doc) => {
-    if (!doc.opened_by_current_user) {
-      return <strong>{content}</strong>;
-    }
-
-    return content;
-  };
-
-  onClick = (id) => () => {
-    // Annoyingly if we make this call in the thread, it won't follow the link. Instead
-    // we use setTimeout to force it to run at a later point.
-    setTimeout(() => this.props.selectCurrentPdfLocally(id), 0);
-  }
-
-  render = () => {
-    const { doc } = this.props;
-
-    // We add a click handler to mark a document as read even if it's opened in a new tab.
-    // This will get fired in the current tab, as the link is followed in a new tab. We
-    // also need to add a mouseUp event since middle clicking doesn't trigger an onClick.
-    // This will not work if someone right clicks and opens in a new tab.
-    return this.boldUnreadContent(
-      <Link
-        onMouseUp={this.onClick(doc.id)}
-        onClick={this.onClick(doc.id)}
-        to={`${this.props.documentPathBase}/${doc.id}`}
-        aria-label={doc.type + (doc.opened_by_current_user ? ' opened' : ' unopened')}>
-        <Highlight>
-          {doc.type}
-        </Highlight>
-      </Link>, doc);
-  }
-}
-
-const mapDocTypeDispatchToProps = (dispatch) => ({
-  ...bindActionCreators({
-    selectCurrentPdfLocally
-  }, dispatch)
-});
-
-DocTypeColumn.propTypes = {
-  doc: PropTypes.object,
-  documentPathBase: PropTypes.string
-};
-
-const ConnectedDocTypeColumn = connect(
-  null, mapDocTypeDispatchToProps
-)(DocTypeColumn);
 
 export const getRowObjects = (documents, annotationsPerDocument, viewingDocumentsOrComments) => {
   return documents.reduce((acc, doc) => {
@@ -295,7 +184,7 @@ class DocumentsTable extends React.Component {
     return [
       {
         cellClass: 'last-read-column',
-        valueFunction: (doc) => <ConnectedLastReadIndicator docId={doc.id} getRef={this.getLastReadIndicatorRef} />
+        valueFunction: (doc) => <LastReadIndicator docId={doc.id} getRef={this.getLastReadIndicatorRef} />
       },
       {
         cellClass: 'categories-column',
@@ -346,7 +235,7 @@ class DocumentsTable extends React.Component {
           onClick={() => this.props.changeSortState('type')}>
           Document Type {this.props.docFilterCriteria.sort.sortBy === 'type' ? sortArrowIcon : notSortedIcon }
         </Button>,
-        valueFunction: (doc) => <ConnectedDocTypeColumn doc={doc}
+        valueFunction: (doc) => <DocTypeColumn doc={doc}
           documentPathBase={this.props.documentPathBase}/>
       },
       {
